// src/api/gen/types/AddressAliasDtoV1.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const addressAliasDtoV1VersionEnum = {
  V1: 'V1',
} as const

export type AddressAliasDtoV1VersionEnumKey =
  (typeof addressAliasDtoV1VersionEnum)[keyof typeof addressAliasDtoV1VersionEnum]

export type AddressAliasDtoV1 = {
  /**
   * @type string | undefined
   */
  version?: AddressAliasDtoV1VersionEnumKey
  /**
   * @type string | undefined
   */
  alias?: string
  /**
   * @type string | undefined
   */
  address?: string
  /**
   * @type string | undefined
   */
  originTxHash?: string
  /**
   * @type integer | undefined, int64
   */
  createdAtBlockHeight?: number
  /**
   * @type string | undefined, date-time
   */
  createdAtTimestamp?: string
}