// src/api/gen/types/AuthorityDtoV1.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const authorityDtoV1VersionEnum = {
  V1: 'V1',
} as const

export type AuthorityDtoV1VersionEnumKey =
  (typeof authorityDtoV1VersionEnum)[keyof typeof authorityDtoV1VersionEnum]

export type AuthorityDtoV1 = {
  /**
   * @type string | undefined
   */
  version?: AuthorityDtoV1VersionEnumKey
  /**
   * @type string | undefined
   */
  address?: string
  /**
   * @type string | undefined
   */
  originTxHash?: string
  /**
   * @type integer | undefined, int64
   */
  createdAtBlockHeight?: number
  /**
   * @type string | undefined, date-time
   */
  createdAtTimestamp?: string
}